--- src/Web/Features/Orders/GetMyOrders.cs
@@ -1,53 +1,0 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-using System.Linq;
-using MediatR;
-using Microsoft.eShopWeb.ApplicationCore.Interfaces;
-using Microsoft.eShopWeb.ApplicationCore.Specifications;
-using Microsoft.eShopWeb.Web.ViewModels;
-using System.Collections.Generic;
-
-namespace Microsoft.eShopWeb.Web.Features.Orders
-{
-    public class GetMyOrders : IRequest<IEnumerable<OrderViewModel>>
-    {
-        public string UserName { get; set; }
-
-        public GetMyOrders(string userName)
-        {
-            UserName = userName;
-        }
-    }
-
-    public class GetMyOrdersHandler : IRequestHandler<GetMyOrders, IEnumerable<OrderViewModel>>
-    {
-        private readonly IOrderRepository _orderRepository;
-
-        public GetMyOrdersHandler(IOrderRepository orderRepository)
-        {
-            _orderRepository = orderRepository;
-        }
-
-        public async Task<IEnumerable<OrderViewModel>> Handle(GetMyOrders request, CancellationToken cancellationToken)
-        {
-            var specification = new CustomerOrdersWithItemsSpecification(request.UserName);
-            var orders = await _orderRepository.ListAsync(specification);
-
-            return orders.Select(o => new OrderViewModel
-                {
-                    OrderDate = o.OrderDate,
-                    OrderItems = o.OrderItems?.Select(oi => new OrderItemViewModel()
-                    {
-                        PictureUrl = oi.ItemOrdered.PictureUri,
-                        ProductId = oi.ItemOrdered.CatalogItemId,
-                        ProductName = oi.ItemOrdered.ProductName,
-                        UnitPrice = oi.UnitPrice,
-                        Units = oi.Units
-                    }).ToList(),
-                    OrderNumber = o.Id,
-                    ShippingAddress = o.ShipToAddress,
-                    Total = o.Total()
-                });
-        }
-    }
-}
