+++ src/Web/Features/Orders/RequestHandlers/GetMyOrdersHandler.cs
@@ -1,0 +1,44 @@
+﻿using MediatR;
+using Microsoft.eShopWeb.ApplicationCore.Interfaces;
+using Microsoft.eShopWeb.ApplicationCore.Specifications;
+using Microsoft.eShopWeb.Web.Features.Orders.Requests;
+using Microsoft.eShopWeb.Web.ViewModels;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Microsoft.eShopWeb.Web.Features.Orders.RequestHandlers
+{
+    public class GetMyOrdersHandler : IRequestHandler<GetMyOrders, IEnumerable<OrderViewModel>>
+    {
+        private readonly IOrderRepository _orderRepository;
+
+        public GetMyOrdersHandler(IOrderRepository orderRepository)
+        {
+            _orderRepository = orderRepository;
+        }
+
+        public async Task<IEnumerable<OrderViewModel>> Handle(GetMyOrders request, CancellationToken cancellationToken)
+        {
+            var specification = new CustomerOrdersWithItemsSpecification(request.UserName);
+            var orders = await _orderRepository.ListAsync(specification);
+
+            return orders.Select(o => new OrderViewModel
+            {
+                OrderDate = o.OrderDate,
+                OrderItems = o.OrderItems?.Select(oi => new OrderItemViewModel()
+                {
+                    PictureUrl = oi.ItemOrdered.PictureUri,
+                    ProductId = oi.ItemOrdered.CatalogItemId,
+                    ProductName = oi.ItemOrdered.ProductName,
+                    UnitPrice = oi.UnitPrice,
+                    Units = oi.Units
+                }).ToList(),
+                OrderNumber = o.Id,
+                ShippingAddress = o.ShipToAddress,
+                Total = o.Total()
+            });
+        }
+    }
+}
